Fix get_args for unparameterized ClassVar on <3.7
In 3.7, `get_args(ClassVar)` (no parameter specified) would return an
empty tuple. In 3.6 it would return `(None,)` because `__type__` is
`None` in that case.

This copies the logic that already existed in `get_last_args` to
normalize that case to an empty tuple.

I've tested this pull request by running the unit tests against Python
2.7, 3.6, and 3.7.