Making the is_new_type check more robust then just checking __supertype__ existence
Before this commit everthing which has an attribute called __supertype__
will be accepted as ‚NewType‘ which is wrong.

Now it is additionally checked that the __qualname__ must be
'NewType.<locals>.new_type', and that the module where the symbol is
defined has the name 'typing'. A user could still fullfill all those
conditions in a custom class, but the probability bacame quite low.

Furtheremore support for python 3.10¹, where NewType is not a function
any more but a class, is added. This check is implemented by
isincance() and should work always correctly.

The symbol ‚NewType‘ (i.e. not the call of it) is now also cosidered as
a NewType by doing a `tp is NewType` check. This is in alignment with
other functions, such as `is_classvar()` which does an `tp is ClassVar`
or `is_union_type()` which do a `tp is Union`.

¹ To be specific, the first RC of python 3.10, i.e. in all bata
versions, NewType is still a function. Please also note that the class
NewType in python 3.10.0rc has a different __qualname__ then the
function NewType before.